6/pyproject.toml
[tool.poetry]
name = "exercise"
version = "0.1.0"
description = "Exercise"
authors = ["exercise"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
